Removed potential typo with no existing evidence
A vehicle tag is listed as having a hard coded change affecting its min and max error angle, vehicle tags have no such error fields at all. I searched various weapon tags to see if it was misplaced info but the only other weapon set to have it error compile as 0.1 is the battle rifle, which is already listed.